+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+- Show comparisons between hours of current month/week vs previous month/week
+
 ## 1.0.0
 
 - Initial Release